Images-Resolution-Enhancement-Usiung-CNN/requirements.txt
ipython==8.9.0
keras==2.11.0
matplotlib==3.6.3
numpy==1.23.4
Pillow==9.4.0
tensorflow==2.11.0
